feat: add dedicated workers and worker pool
This addition enhances our scheduler by adding dedicated workers and a
worker pool to manage all workers.

Another minor addition is the improvement of the workflows, their file
extensions have been renamed, and the workflow now includes a check with
clippy as well as a audit check at the end.

Fixes #15, #17, #18, #20